Simplify data passing with MVC pattern
