--- VideoPlayer/VideoPlayer/Utility/Box.swift
@@ -1,29 +1,0 @@
-//
-//  Box.swift
-//  VideoPlayer
-//
-//  Created by Zheng-Yuan Yu on 2022/1/3.
-//
-
-import Foundation
-
-final class Box<T> {
-
-  typealias Listener = (T) -> Void
-  var listener: Listener?
-
-  var value: T {
-    didSet {
-      listener?(value)
-    }
-  }
-
-  init(_ value: T) {
-    self.value = value
-  }
-
-  func bind(listener: Listener?) {
-    self.listener = listener
-    listener?(value)
-  }
-}
